% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paulifier.R
\name{paulify}
\alias{paulify}
\title{Take a character vector and replace words Paul doesn't like with
words he does like}
\usage{
paulify(string)
}
\arguments{
\item{string}{A character vector with bad words in it}
}
\value{
A character vector with no bad words
}
\description{
Take a character vector and replace words Paul doesn't like with
words he does like
}
\examples{
paulify("Tom's teeth were impacted")
}
